Renamed readers/bz2 to inputs/bz2
I want new input sources to be easy to add to wikiporter, so it makes
more sense to place them all inside an 'inputs' directory than it does
to put them inside of a 'readers' directory, as every input might not be
a reader.

I have updated the namespaces accordingly.