Added postgresql output component and misc
I have written an initial postgresql output component which uses
hikari-cp to manage a connection pool. It currently has a function to
insert a list of maps, though the table it inserts into is hard coded,
which I'll be changing. It also has a function to easily delete
everything in a table, which I wrote because it's helpful during
testing.

I have also added a sample data set, which is very small, as
wikisample.short.xml.bz2, and a file holding some misc. functions and
info on how to call the system into being from the repl.