Initial cli work complete
The system can now be called from the command line, e.g. with a jar or
from lein run. Passing the -h or --help flag to Wikiporter will make it
print out the cli options and quit.

Currently the system will only do a shallow import of the pages,
supporting a small subset of the tags/columns.